requirements.txt
Flask==2.3.3
Flask-Limiter==3.5.0
bcrypt==4.0.1
PyJWT==2.8.0
qrcode==7.4.2
Pillow==10.0.0
pyotp==2.9.0
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.31.0 